chore(md): removed old ui code
