--- packages/core/src/managed/ColumnHeader.less
@@ -1,23 +1,0 @@
-.ColumnHeader {
-  background-color: rgb(100, 140, 175);
-  display: flex;
-  height: 32px;
-  justify-content: space-between;
-  align-items: center;
-  color: var(--color-white);
-  font-size: 14px;
-  border-radius: 2px;
-  z-index: var(--layer-low);
-  flex: none;
-
-  .column-header-icon {
-    flex: 0 0 24px;
-    margin-left: 12px;
-  }
-
-  .column-header-text {
-    flex: 1 1 1px;
-    text-align: center;
-    margin: 0 12px;
-  }
-}
