--- packages/core/src/managed/EnvironmentBadge.less
@@ -1,20 +1,0 @@
-.EnvironmentBadge {
-  border-radius: 2px;
-  text-transform: uppercase;
-  color: var(--color-white);
-  background-color: #53809f;
-  line-height: normal;
-
-  &.small,
-  &.extraSmall {
-    font-size: 12px;
-  }
-
-  &.medium {
-    font-size: 14px;
-  }
-
-  .critical-symbol {
-    vertical-align: middle;
-  }
-}
