--- packages/core/src/managed/EnvironmentsHeader.less
@@ -1,29 +1,0 @@
-.EnvironmentsHeader {
-  background-color: var(--color-white);
-  border-radius: 3px;
-  box-shadow: 0 1px 3px 0 rgba(0, 0, 0, 0.12);
-  margin: 0 3px 16px 3px;
-
-  // dropdown-toggle class is from the bootstrap library and needed here to increase the specificity of our rule
-  // to avoid bootstrap changing the background and text colors to default value.
-  .dropdown-toggle.dropdown-toggle-btn,
-  .dropdown-toggle.dropdown-toggle-btn:hover,
-  .open > .dropdown-toggle.dropdown-toggle-btn,
-  .open > .dropdown-toggle.dropdown-toggle-btn:hover,
-  .open > .dropdown-toggle.dropdown-toggle-btn:focus {
-    background-color: rgb(90, 95, 180);
-    color: var(--color-white);
-  }
-
-  .dropdown-menu > li {
-    cursor: pointer;
-
-    &:hover {
-      background-color: var(--color-alabaster);
-    }
-
-    a {
-      padding: 8px 20px;
-    }
-  }
-}
