--- packages/core/src/managed/StatusCard.less
@@ -1,43 +1,0 @@
-.StatusCard {
-  font-size: 13px;
-  color: var(--color-icon-neutral);
-
-  &.active {
-    color: var(--color-charcoal);
-  }
-
-  // Need to override this, because the 'future' status should always appear inactive
-  &.status-card-future.active {
-    color: var(--color-icon-neutral);
-  }
-
-  &.with-background {
-    &.status-card-neutral {
-      background-color: var(--color-status-neutral-light);
-    }
-
-    &.status-card-info {
-      background-color: var(--color-status-info-light);
-    }
-
-    &.status-card-progress {
-      background-color: var(--color-status-progress-light);
-    }
-
-    &.status-card-success {
-      background-color: var(--color-status-success-light);
-    }
-
-    &.status-card-warning {
-      background-color: var(--color-status-warning-light);
-    }
-
-    &.status-card-error {
-      background-color: var(--color-status-error-light);
-    }
-
-    &.status-card-archived {
-      background-color: var(--color-status-archived-light);
-    }
-  }
-}
