--- packages/core/src/managed/artifactDetail/ArtifactDetail.less
@@ -1,28 +1,0 @@
-.ArtifactDetail {
-  background-color: var(--color-white);
-  padding: 16px 16px 64px 16px;
-  overflow-y: scroll;
-
-  .detail-section-left {
-    flex: 0 0 280px;
-    font-style: italic;
-    font-size: 14px;
-    color: var(--color-charcoal);
-  }
-
-  .detail-section-right {
-    max-width: 650px;
-    color: var(--color-charcoal);
-
-    .metadata-label {
-      width: 100px;
-      color: var(--color-black);
-    }
-  }
-
-  .resources-section {
-    margin-top: 8px;
-    padding-top: 8px;
-    border-top: 1px solid var(--color-accent-g3);
-  }
-}
