--- packages/core/src/managed/artifactDetail/ArtifactDetailHeader.less
@@ -1,41 +1,0 @@
-.ArtifactDetailHeader {
-  position: sticky;
-  top: 0;
-  height: 64px;
-  z-index: 1;
-  background-color: rgb(100, 140, 175);
-  color: var(--color-white);
-  font-size: 14px;
-  padding: 0 8px 0 16px;
-  border-radius: 2px 2px 0 0;
-
-  .header-section-left {
-    flex: 0 0 112px;
-  }
-
-  .header-version-pill {
-    margin-left: 16px;
-    background-color: var(--color-white);
-    color: rgb(100, 140, 175);
-    border-radius: 24px;
-    line-height: 24px;
-    height: 24px;
-    font-size: 16px;
-    flex: 0 0 54px;
-    padding: 0 12px 0 10px;
-    text-align: center;
-    white-space: nowrap;
-  }
-
-  .header-close {
-    cursor: pointer;
-    width: 48px;
-    height: 48px;
-    margin-left: 72px;
-    border-radius: 2px;
-  }
-
-  .header-close:hover {
-    background-image: linear-gradient(to bottom, rgba(160, 180, 220, 0.2), rgba(160, 180, 220, 0.2));
-  }
-}
