--- packages/core/src/managed/artifactDetail/PreDeploymentRow.less
@@ -1,17 +1,0 @@
-.PreDeploymentRow {
-  max-width: 100%;
-
-  .srow {
-    display: flex;
-    align-items: center;
-    box-shadow: inset 0 -1px 0 0 var(--color-nobel);
-    height: 44px;
-    font-size: 16px;
-    color: var(--color-nobel);
-    padding-left: 8px;
-  }
-
-  .titleColumn {
-    flex-basis: 50%;
-  }
-}
