--- packages/core/src/managed/artifactsList/ArtifactRow.less
@@ -1,147 +1,0 @@
-.ArtifactRow {
-  height: 94px;
-  display: block;
-  position: relative;
-  cursor: pointer;
-  margin-bottom: 2px;
-  border-radius: 2px;
-  transition: background-color 150ms ease-in-out;
-
-  &:first-of-type {
-    margin-top: var(--m-spacing);
-  }
-
-  &:last-of-type {
-    margin-bottom: var(--m-spacing);
-  }
-
-  &:not(.selected):hover {
-    background-color: #e8eaf2;
-  }
-
-  &.selected {
-    background-color: #dbe5eb;
-  }
-
-  .row-content {
-    height: 100%;
-  }
-
-  .row-middle-section {
-    width: 100%;
-  }
-
-  .version-title {
-    max-width: 100%;
-    overflow: hidden;
-  }
-
-  .version-name {
-    font-size: 13px;
-    color: var(--color-icon-neutral);
-    text-overflow: ellipsis;
-    white-space: nowrap;
-    overflow: hidden;
-  }
-
-  .version-secondary-summary {
-    font-size: 13px;
-    font-style: italic;
-    color: var(--color-icon-neutral);
-    text-overflow: ellipsis;
-    white-space: nowrap;
-    overflow: hidden;
-    margin-left: 20px;
-    opacity: 1;
-    transition: opacity 150ms ease-in-out;
-  }
-
-  &:hover .version-secondary-summary {
-    opacity: 0;
-  }
-
-  .environment-stages {
-    display: flex;
-    position: absolute;
-    height: 8px;
-    bottom: 8px;
-    left: 12px;
-    right: 12px;
-    z-index: 1;
-    transition: height 150ms ease-in-out;
-  }
-
-  .environment-stage {
-    display: flex;
-    flex: 1 1 auto;
-    font-size: 12px;
-    color: rgba(0, 0, 0, 0.5);
-    border: 1px solid rgba(0, 0, 0, 0.5);
-    margin-right: 2px;
-    text-transform: uppercase;
-    transition: background-color 150ms ease-in-out;
-
-    &.current {
-      background-color: var(--color-status-success);
-    }
-
-    &.approved {
-      background-color: var(--color-status-info);
-    }
-
-    &.deploying {
-      background-color: var(--color-status-progress);
-    }
-
-    &.pending,
-    &.skipped {
-      background-color: var(--color-alabaster);
-    }
-
-    &.previous {
-      background-color: #dbcec8;
-    }
-
-    &.vetoed {
-      background-color: var(--color-status-error);
-    }
-  }
-
-  .environment-stage:last-child {
-    margin-right: 0;
-  }
-
-  .environment-stage-name {
-    opacity: 0;
-    user-select: none;
-    pointer-events: none;
-    text-overflow: ellipsis;
-    overflow: hidden;
-    white-space: nowrap;
-    transition: opacity 150ms ease-in-out;
-  }
-
-  &:not(.selected):hover {
-    .environment-stage.pending,
-    .environment-stage.skipped {
-      background-color: #e8eaf2;
-    }
-  }
-
-  &.selected {
-    .environment-stage.pending,
-    .environment-stage.skipped {
-      background-color: #dbe5eb;
-    }
-  }
-
-  &:hover {
-    .environment-stages {
-      height: 21px;
-    }
-
-    .environment-stage-name {
-      opacity: 1;
-    }
-  }
-}
