--- packages/core/src/managed/artifactsList/StatusBubbleStack.less
@@ -1,22 +1,0 @@
-.StatusBubbleStack {
-  // We want to create a new stacking context so that all
-  // the bubbles go in front/behind other elements as a unit
-  // vs. each individual bubble having its own stacking.
-  transform: translate(0px, 0px);
-}
-
-.StatusBubbleContainer {
-  background-color: var(--color-alabaster);
-  border-radius: 50%;
-  padding: 2px;
-  margin-left: -8px;
-}
-
-.HiddenStatusCount {
-  border-radius: 50%;
-  background-color: var(--color-status-neutral);
-  color: var(--color-white);
-  margin-left: -4px;
-  margin-right: 6px;
-  padding: 4px 8px;
-}
