--- packages/core/src/managed/constraints/ConstraintCard.less
@@ -1,3 +1,0 @@
-.action-error-message {
-  color: #be0000;
-}
