--- packages/core/src/managed/environment/EnvironmentRow.less
@@ -1,50 +1,0 @@
-.EnvironmentRow {
-  max-width: 100%;
-
-  .srow {
-    display: flex;
-    align-items: center;
-    box-shadow: inset 0 -1px 0 0 #53809f;
-    height: 44px;
-    font-size: 14px;
-  }
-
-  .srow:hover .select {
-    visibility: visible;
-  }
-
-  .clickableArea {
-    flex: 1 1 auto;
-    height: 100%;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding-left: 8px;
-  }
-
-  .titleColumn {
-    flex-basis: 50%;
-  }
-
-  .select {
-    margin-right: -40px;
-    visibility: hidden;
-  }
-
-  .select,
-  .expand {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 44px;
-    height: 100%;
-    user-select: none;
-  }
-
-  //   .clickableArea:hover,
-  //   .select:hover,
-  .expand:hover {
-    background-image: linear-gradient(to bottom, rgba(160, 180, 220, 0.2), rgba(160, 180, 220, 0.2));
-    cursor: pointer;
-  }
-}
