+++ packages/core/src/managed/managed.dataSource.ts
@@ -34,18 +34,6 @@ module(MANAGED_RESOURCES_DATA_SOURCE, []).run([
       application.securityGroups.ready().then(() => addManagedResourceMetadataToSecurityGroups(application), noop);
     };
 
-    const loadEnvironments = (application: Application) => {
-      return ManagedReader.getEnvironmentsSummary(application.name);
-    };
-
-    const addEnvironments = (
-      application: Application,
-      data: IManagedApplicationSummary<'resources' | 'artifacts' | 'environments'>,
-    ) => {
-      application.isManagementPaused = data.applicationPaused;
-      return $q.when(data);
-    };
-
     ApplicationDataSourceRegistry.registerDataSource({
       key: 'managedResources',
       visible: false,
@@ -65,8 +53,6 @@ module(MANAGED_RESOURCES_DATA_SOURCE, []).run([
       icon: 'fa fa-fw fa-xs fa-code-branch',
       iconName: 'spEnvironments',
       description: 'Artifacts and environments managed by Spinnaker',
-      loader: loadEnvironments,
-      onLoad: addEnvironments,
       defaultData: {
         applicationPaused: false,
         hasManagedResources: false,
