+++ packages/core/src/managed/utils/ActionModal.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { Button } from '../Button';
+import { Button } from './Button';
 import { useLogEvent } from './logging';
 import {
   FormikFormField,
